backend/requirements.txt
aioboto3==15.5.0
argon2-cffi==25.1.0
async-lru==2.0.5
boto3==1.40.61
build==1.2.2.post1
docopt==0.6.2
fastapi==0.115.13
google-api-python-client==2.174.0
google-genai==1.21.1
greenlet==3.2.3
grpcio-status==1.71.2
httptools==0.6.4
ipython==8.12.3
mistralai==1.9.10
motor==3.7.1
nbconvert==7.16.6
nltk==3.9.1
oauth2client==4.1.3
openai==1.97.1
pdf2image==1.17.0
pip==25.2
poppler-utils==0.1.0
proto-plus==1.26.1
psycopg==3.2.9
psycopg-binary==3.2.9
pydantic-settings==2.10.0
PyJWT==2.10.1
pyLDAvis==3.4.1
pypdf==5.6.1
PyPDF2==3.0.1
python-multipart==0.0.20
spacy==3.8.7
SQLAlchemy==2.0.41
tinycss2==1.4.0
uuid==1.30
uvicorn==0.34.3
uvloop==0.21.0
watchfiles==1.1.0
wheel==0.45.1
yarg==0.1.9